# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API-сервер для управления заявками и стратегиями восстановления данных.",
        "title": "API Системы восстановления",
        "contact": {
            "name": "API Support",
            "email": "support@recovery-api.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/recovery_requests": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает список заявок. Для модератора - все, для пользователя - только его.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "requests"
                ],
                "summary": "Получить список заявок (авторизованные пользователи)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Фильтр по статусу заявки (formed, completed, rejected)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Фильтр по дате 'от' (формат YYYY-MM-DD)",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Фильтр по дате 'до' (формат YYYY-MM-DD)",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ds.RequestDTO"
                            }
                        }
                    },
                    "401": {
                        "description": "Необходима авторизация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/recovery_requests/cart": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает ID черновика текущего пользователя и количество стратегий в нем.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "requests"
                ],
                "summary": "Получить информацию для иконки корзины (авторизованные пользователи)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ds.CartBadgeDTO"
                        }
                    },
                    "401": {
                        "description": "Необходима авторизация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/recovery_requests/draft/strategies/{strategy_id}": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Находит или создает черновик заявки для текущего пользователя и добавляет в него стратегию.",
                "tags": [
                    "strategies"
                ],
                "summary": "Добавить стратегию в черновик заявки (авторизованные пользователи)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID стратегии для добавления",
                        "name": "strategy_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "401": {
                        "description": "Необходима авторизация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/recovery_requests/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает полную информацию о заявке, включая привязанные стратегии.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "requests"
                ],
                "summary": "Получить одну заявку по ID (авторизованные пользователи)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ds.RequestDTO"
                        }
                    },
                    "401": {
                        "description": "Необходима авторизация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Заявка не найдена",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Позволяет пользователю обновить поля своей заявки (уровень навыков, пропускная способность сети и т.д.).",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "requests"
                ],
                "summary": "Обновить детали заявки (авторизованные пользователи)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные для обновления",
                        "name": "updateData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.UpdateRequestDetailsRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Необходима авторизация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Логически удаляет заявку, переводя ее в статус \"удалена\". Доступно только для создателя и только для черновиков.",
                "tags": [
                    "requests"
                ],
                "summary": "Удалить заявку (черновик) (авторизованные пользователи)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Необходима авторизация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/recovery_requests/{id}/form": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Переводит заявку из статуса \"черновик\" в \"сформирована\".",
                "tags": [
                    "requests"
                ],
                "summary": "Сформировать заявку (авторизованные пользователи)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки (черновика)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Необходима авторизация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/recovery_requests/{id}/resolve": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Модератор завершает (с расчетом) или отклоняет заявку.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "requests"
                ],
                "summary": "Завершить или отклонить заявку (модератор)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Действие: 'complete' или 'reject'",
                        "name": "action",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.ResolveRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Необходима авторизация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/recovery_requests/{id}/strategies/{strategy_id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Изменяет дополнительные данные (например, объем данных) для конкретной стратегии в рамках одной заявки.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "m-m"
                ],
                "summary": "Обновить данные стратегии в заявке (авторизованные пользователи)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID стратегии",
                        "name": "strategy_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Новые данные",
                        "name": "updateData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.UpdateRequestStrategyRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Необходима авторизация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Удаляет связь между заявкой и стратегией.",
                "tags": [
                    "m-m"
                ],
                "summary": "Удалить стратегию из заявки (авторизованные пользователи)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заявки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID стратегии",
                        "name": "strategy_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Необходима авторизация",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "ds.CartBadgeDTO": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "request_id": {
                    "type": "integer"
                }
            }
        },
...
        "ds.RequestDTO": {
            "type": "object",
            "properties": {
                "calculated_recovery_time_hours": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "creator_username": {
                    "type": "string"
                },
                "documentation_quality": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "it_skill_level": {
                    "type": "string"
                },
                "moderator_username": {
                    "type": "string"
                },
                "network_bandwidth_mbps": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "strategies": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ds.StrategyDTO"
                    }
                }
            }
        },
        "ds.ResolveRequest": {
            "type": "object",
            "required": [
                "action"
            ],
            "properties": {
                "action": {
                    "description": "\"complete\" or \"reject\"",
                    "type": "string"
                }
            }
        },
        "ds.StrategyDTO": {
            "type": "object",
            "properties": {
                "base_recovery_hours": {
                    "type": "number"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "ds.UpdateRequestDetailsRequest": {
            "type": "object",
            "properties": {
                "documentation_quality": {
                    "type": "string"
                },
                "it_skill_level": {
                    "type": "string"
                },
                "network_bandwidth_mbps": {
                    "type": "integer"
                }
            }
        },
        "ds.UpdateRequestStrategyRequest": {
            "type": "object",
            "required": [
                "data_to_recover_gb"
            ],
            "properties": {
                "data_to_recover_gb": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}